refactor: centralize DOM setup for tests
