+++ coins.test.js
@@ -1,63 +1,11 @@
 import test from 'node:test';
 import assert from 'node:assert';
-import { Game } from './src/game.js';
+import { createStubGame } from './testHelpers.js';
 import { Obstacle } from './src/obstacle.js';
 import { Level2 } from './src/levels/level2.js';
 import { LEVEL_UP_SCORE } from './src/config.js';
 
 const FRAME = 1 / 60;
-
-function createStubGame() {
-  const noop = () => {};
-  const ctx = {
-    clearRect: noop,
-    fillRect: noop,
-    beginPath: noop,
-    moveTo: noop,
-    lineTo: noop,
-    fill: noop,
-    arc: noop,
-    stroke: noop,
-    fillText: noop,
-    measureText: () => ({ width: 0 }),
-  };
-  const canvas = { width: 800, height: 200, getContext: () => ctx };
-  const overlay = { classList: { add: noop, remove: noop } };
-  const overlayContent = {};
-  const overlayButton = {};
-
-  global.document = {
-    getElementById: (id) => {
-      switch (id) {
-        case 'game':
-          return canvas;
-        case 'overlay':
-          return overlay;
-        case 'overlay-content':
-          return overlayContent;
-        case 'overlay-button':
-          return overlayButton;
-        default:
-          return null;
-      }
-    },
-    addEventListener: noop,
-    removeEventListener: noop,
-  };
-
-  global.window = {
-    innerWidth: 800,
-    location: { search: '' },
-    addEventListener: noop,
-    removeEventListener: noop,
-    requestAnimationFrame: noop,
-  };
-
-  const game = new Game(canvas);
-  game.showOverlay = noop;
-  game.gamePaused = false;
-  return game;
-}
 
 test('awards coin for passed obstacle and preserves coins in level 2', () => {
   const game = createStubGame();
